Create Frogger (Konami Bugfixed).mra
Hi,

I'm not sure how to contribute this but there is a patch to the konami frogger romfiles that fixes some bugs in frogger over at:

[http://www.ukvac.com/forum/frogger-bugs-descriptions-and-fixes_topic381374.html](url)

These issues appear to be present in the current mister sega rom set too. 

Anyways, I took the file from that forum and made up a .mra that applies the patch to the mame konami roms. I was hoping that somehow this could be added to the distro...

Perhaps this sound go in the alternatives directory?

Jay